Extract BoxShadow from box_painter.dart
This is the last class other than BoxDecoration in this file, so by
extracting it we finally have all the parts of BoxDecoration
extracted.
